airbyte-integrations/connectors/source-quickbooks-drivepoint/pyproject.toml
[project]
name = "source-quickbooks-drivepoint"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrey Gubarevich",email = "andrey@bainbridgegrowth.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "airbyte-cdk (>=6.60.14,<7.0.0)",
    "pytest (>=8)",
    "pytest-lazy-fixtures (>= 1.1.4)",
    "requests-mock (>=1.12.1)",
    "freezegun (>=1.5.2)",
    "pendulum (>=3.1.0)",
    "dateparser (>=1.2.1)",
    "pytest-mock (>=3.14.1)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "source_quickbooks_drivepoint"

[tool.poetry.scripts]
source-quickbooks-drivepoint = "source_quickbooks_drivepoint.run:run"

[tool.poetry.group.dev.dependencies]
requests-mock = "^1"
pytest = "^8.0.0"
pytest-mock = "^3.6.1"
freezegun = "^1.5.5"